fix: 이메일 인증 로직 수정 및 관련 코드 정리
